[#29]Feat: 게임수, 평균, 최고, 최저점수 조회 구현

+++ src/main/java/capstone/SportyUp/SportyUp_Server/web/DTO/AnalyzeDTO/AnalyzeResponseDTO.java
@@ -15,8 +15,8 @@ public class AnalyzeResponseDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class ChartDTO{
-        Integer gameCount;  //전체 게임 수
-        Integer averageScore; //전체 평균 점수
+        Long gameCount;  //전체 게임 수
+        Double averageScore; //전체 평균 점수
         Integer highScore;  //최고 점수
         Integer lowScore;   //최저 점수
         List<DateScoreDTO> dateScores;  //날짜별 점수 리스트
